4,5,6,7 Nat Gateway...
The Subnet called “Public” needs to have a NAT Gateway deployed in it
This will require you to allocate an Elastic IP that you can then use to assign it to the NAT Gateway.
The Public Subnet needs to have the MapPublicIpOnLaunch property set to true. Use this reference for help.
The Private Subnet needs to have the MapPublicIpOnLaunch property set to false.
Both subnets need to be /24 in size